pyproject.toml
[project]
name = "paste-bar-code"
description = ""
version = "0.0.1"
requires-python = ">=3.12,<3.13"
dependencies = [
    "cv2-enumerate-cameras>=1.3.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "pyside6==6.9.1",
    "pyzbar>=0.1.9",
]

[tool.ruff]
line-length = 100
select = ["I"]
